source/Test.Stateless.WorkflowEngine.WebConsole/Modules/StoreModuleTest.cs
using AutoMapper;
using Encryption;
using Nancy;
using Nancy.Authentication.Forms;
using Nancy.Bootstrapper;
using Nancy.Responses.Negotiation;
using Nancy.Testing;
using Nancy.ViewEngines.Razor;
using Newtonsoft.Json;
using NSubstitute;
using NUnit.Framework;
using Stateless.WorkflowEngine.WebConsole.BLL.Data.Models;
using Stateless.WorkflowEngine.WebConsole.BLL.Data.Stores;
using Stateless.WorkflowEngine.WebConsole.BLL.Models;
using Stateless.WorkflowEngine.WebConsole.BLL.Security;
using Stateless.WorkflowEngine.WebConsole.BLL.Services;
using Stateless.WorkflowEngine.WebConsole.BLL.Validators;
using Stateless.WorkflowEngine.WebConsole.Modules;
using Stateless.WorkflowEngine.WebConsole.Navigation;
using Stateless.WorkflowEngine.WebConsole.ViewModels;
using Stateless.WorkflowEngine.WebConsole.ViewModels.Login;
using Stateless.WorkflowEngine.WebConsole.ViewModels.Store;
using System;
using System.Collections.Generic;
using System.Configuration;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Test.Stateless.WorkflowEngine.WebConsole.Modules
{
    [TestFixture]
    public class StoreModuleTest
    {
        private StoreModule _storeModule;
        private IUserStore _userStore;
        private IWorkflowInfoService _workflowInfoService;

        [SetUp]
        public void UserModuleTest_SetUp()
        {
            _userStore = Substitute.For<IUserStore>();
            _workflowInfoService = Substitute.For<IWorkflowInfoService>();

            _storeModule = new StoreModule(_userStore, null);

        }

        #region Default Tests

        [Test]
        public void Default_NoConnectionFound_ReturnsNotFoundResponse()
        {
            // setup
            var bootstrapper = this.ConfigureBootstrapperAndUser(false);
            var browser = new Browser(bootstrapper);
            var connectionId = Guid.NewGuid();

            List<UserModel> users = ConfigureUsers(bootstrapper);
            UserModel currentUser = users[0];
            _userStore.GetUser(currentUser.UserName).Returns(currentUser);
            currentUser.Connections = new List<ConnectionModel>();

            // execute
            var response = browser.Get(Actions.Store.Default, (with) =>
            {
                with.HttpRequest();
                with.FormsAuth(bootstrapper.CurrentUser.Id, new Nancy.Authentication.Forms.FormsAuthenticationConfiguration());
                with.Query("id", connectionId.ToString());
            });

            // assert
            Assert.AreEqual(HttpStatusCode.NotFound, response.StatusCode);
        }

        [Test]
        public void Default_ConnectionFound_SetsModelAndReturnsView()
        {
            // setup
            var bootstrapper = this.ConfigureBootstrapperAndUser(false);
            var browser = new Browser(bootstrapper);
            var connectionId = Guid.NewGuid();

            ConnectionModel connection = new ConnectionModel()
            {
                Id = connectionId,
                Host = "myserver"
            };

            List<UserModel> users = ConfigureUsers(bootstrapper);
            UserModel currentUser = users[0];
            _userStore.GetUser(currentUser.UserName).Returns(currentUser);
            currentUser.Connections = new List<ConnectionModel>() { connection };

            Assert.AreEqual(1, currentUser.Connections.Count);

            // execute
            var response = browser.Get(Actions.Store.Default, (with) =>
            {
                with.HttpRequest();
                with.FormsAuth(bootstrapper.CurrentUser.Id, new Nancy.Authentication.Forms.FormsAuthenticationConfiguration());
                with.Query("id", connectionId.ToString());
            });

            // assert
            Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
            Assert.AreEqual(response.ContentType, "text/html");
            
            // string responseBody = response.Body.AsString();

            response.Body["title"]
                .ShouldExistOnce()
                .And.ShouldContain(connection.Host);
        }
        #endregion

        #region List Tests

        [Test]
        public void List_NoConnectionFound_ReturnsNotFoundResponse()
        {
            // setup
            var bootstrapper = this.ConfigureBootstrapperAndUser(false);
            var browser = new Browser(bootstrapper);
            var connectionId = Guid.NewGuid();

            List<UserModel> users = ConfigureUsers(bootstrapper);
            UserModel currentUser = users[0];
            _userStore.GetUser(currentUser.UserName).Returns(currentUser);
            currentUser.Connections = new List<ConnectionModel>();

            // execute
            var response = browser.Post(Actions.Store.List, (with) =>
            {
                with.HttpRequest();
                with.FormsAuth(bootstrapper.CurrentUser.Id, new Nancy.Authentication.Forms.FormsAuthenticationConfiguration());
                with.FormValue("id", connectionId.ToString());
            });

            // assert
            Assert.AreEqual(HttpStatusCode.NotFound, response.StatusCode);

            _workflowInfoService.DidNotReceive().GetIncompleteWorkflows(Arg.Any<ConnectionModel>(), Arg.Any<int>());
        }

        [Test]
        public void List_ConnectionFound_SetsModelAndReturnsView()
        {
            // setup
            var bootstrapper = this.ConfigureBootstrapperAndUser(false);
            var browser = new Browser(bootstrapper);
            var connectionId = Guid.NewGuid();
            int count = new Random().Next(5, 20);

            ConnectionModel connection = new ConnectionModel()
            {
                Id = connectionId,
                Host = "myserver"
            };

            List<UserModel> users = ConfigureUsers(bootstrapper);
            UserModel currentUser = users[0];
            _userStore.GetUser(currentUser.UserName).Returns(currentUser);
            currentUser.Connections = new List<ConnectionModel>() { connection };

            List<UIWorkflow> workflows = new List<UIWorkflow>();
            for (int i = 0; i < count; i++)
            {
                UIWorkflow wf = new UIWorkflow();
                wf.Id = Guid.NewGuid();
                wf.WorkflowType = typeof(UIWorkflow).FullName;
                workflows.Add(wf);
            }
            _workflowInfoService.GetIncompleteWorkflows(connection, 50).Returns(workflows);

            // execute
            var response = browser.Post(Actions.Store.List, (with) =>
            {
                with.HttpRequest();
                with.FormsAuth(bootstrapper.CurrentUser.Id, new Nancy.Authentication.Forms.FormsAuthenticationConfiguration());
                with.FormValue("id", connectionId.ToString());
            });

            // assert
            Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
            Assert.AreEqual(response.ContentType, "text/html");

            // SUCKS - for some reason this view throws a reference exception trying to load 
            // Stateless.WorkflowEngine which is handled in the app with the config file.  Not sure 
            // how to do this programatically so leaving this test for now...
            // Adding the assembly to the test RazorConfiguration didn't help and broke other tests....

            //string responseBody = response.Body.AsString();
            //response.Body["table"].ShouldExistOnce();

            //response.Body["td"]
            //    .ShouldExistExactly(count);
        }
        #endregion
        #region Private Methods

        private ModuleTestBootstrapper ConfigureBootstrapperAndUser(bool configureUsers = true)
        {
            var bootstrapper = new ModuleTestBootstrapper();
            bootstrapper.Login();
            bootstrapper.ConfigureRequestContainerCallback = (container) =>
            {
                container.Register<IUserStore>(_userStore);
                container.Register<IWorkflowInfoService>(_workflowInfoService);
            };

            if (configureUsers)
            {
                ConfigureUsers(bootstrapper);
            }
            return bootstrapper;
        }

        private List<UserModel> ConfigureUsers(ModuleTestBootstrapper bootstrapper)
        {
            // set up the logged in user
            UserModel user = new UserModel()
            {
                Id = bootstrapper.CurrentUser.Id,
                UserName = bootstrapper.CurrentUser.UserName,
                Role = bootstrapper.CurrentUser.Claims.First()
            };
            List<UserModel> users = new List<UserModel>() { user };
            _userStore.Users.Returns(users);
            return users;
        }

        #endregion


    }
}
